Rename vulas_diff dataset to vulas_diff_string
