+++ models/diff_string/char_cnn/__main__.py
@@ -10,7 +10,7 @@ from models.diff_string.char_cnn.model import CharCNN
 from common.evaluation import EvaluatorFactory
 from common.train import TrainerFactory
 from datasets.aapd import AAPDCharQuantized as AAPD
-from datasets.vulas_diff import VulasDiffCharQuantized as IMDB
+from datasets.vulas_diff_string import VulasDiffCharQuantized as IMDB
 from datasets.reuters import ReutersCharQuantized as Reuters
 from datasets.yelp2014 import Yelp2014CharQuantized as Yelp2014
